requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-PyMongo==2.3.0
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1
gunicorn==21.2.0
pymongo==4.6.1
dnspython==2.6.0
alembic==1.16.5
SQLAlchemy==2.0.29
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-dotenv==1.0.0
PyJWT==2.8.0
cloudinary>=1.39.0
flask-socketio